Crud grupo e novos endpoints no crud usuário

+++ src/main/java/com/ifconnect/ifconnectbackend/grupo/GrupoController.java
@@ -1,7 +1,8 @@
-package com.ifconnect.ifconnectbackend.curso;
+package com.ifconnect.ifconnectbackend.grupo;
 
 import com.ifconnect.ifconnectbackend.exception.ErrorDetails;
-import com.ifconnect.ifconnectbackend.models.Curso;
+import com.ifconnect.ifconnectbackend.models.Grupo;
+import com.ifconnect.ifconnectbackend.models.Usuario;
 import com.ifconnect.ifconnectbackend.models.modelvo.SearchFilter;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -13,7 +14,8 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.context.SecurityContext;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -25,13 +27,13 @@ import static org.springframework.http.ResponseEntity.*;
 import static org.springframework.web.servlet.support.ServletUriComponentsBuilder.fromCurrentRequestUri;
 
 @RestController
-@RequestMapping("/api/v1/cursos")
+@RequestMapping("/api/v1/grupos")
 @RequiredArgsConstructor
-public class CursoController {
+public class GrupoController {
 
-    private final CursoService service;
+    private final GrupoService service;
 
-    @Operation(summary = " Save cursos", description = "Return save cursos")
+    @Operation(summary = " Save grupos", description = "Return save grupos")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Request Ok"),
             @ApiResponse(responseCode = "401", description = "Not authenticated agent (missing or invalid credentials)"),
@@ -41,12 +43,11 @@ public class CursoController {
                     content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDetails.class))})})
     @ResponseStatus(value = HttpStatus.CREATED)
     @PostMapping
-    @PreAuthorize("hasRole('ROLE_MANAGER')")
-    public ResponseEntity<Curso> save(@RequestBody @Valid Curso entity) {
+    public ResponseEntity<Grupo> save(@RequestBody @Valid Grupo entity) {
         return created(fromCurrentRequestUri().path(service.saveOrUpdate(entity).getId().toString()).build().toUri()).body(entity);
     }
 
-    @Operation(summary = "Cursos by id.", description = "Return the cursos by id.")
+    @Operation(summary = "Grupos by id.", description = "Return the grupos by id.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Request Ok"),
             @ApiResponse(responseCode = "401", description = "Not authenticated agent (missing or invalid credentials)"),
@@ -56,10 +57,11 @@ public class CursoController {
                     content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDetails.class))})})
     @ResponseStatus(value = HttpStatus.OK)
     @GetMapping("{id}")
-    public ResponseEntity<Curso> getOne(@PathVariable(value = "id") Integer id) {
+    public ResponseEntity<Grupo> getOne(@PathVariable(value = "id") Integer id) {
         return ok().body(service.findById(id));
     }
-    @Operation(summary = "List of cursos.", description = "List of cursos.")
+
+    @Operation(summary = "List of grupos.", description = "List of grupos.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Request Ok"),
             @ApiResponse(responseCode = "401", description = "Not authenticated agent (missing or invalid credentials)"),
@@ -69,11 +71,25 @@ public class CursoController {
                     content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDetails.class))})})
     @ResponseStatus(value = HttpStatus.OK)
     @GetMapping("/get-all")
-    public ResponseEntity<List<Curso>> getAll() {
+    public ResponseEntity<List<Grupo>> getAll() {
         return ok().body(service.findAll());
     }
 
-    @Operation(summary = "The list of cursos Pageable.", description = "Returns the list of entities.")
+    @Operation(summary = "List of grupos of an admin.", description = "List of grupos of an admin.")
+    @ApiResponses(value = {
+            @ApiResponse(responseCode = "200", description = "Request Ok"),
+            @ApiResponse(responseCode = "401", description = "Not authenticated agent (missing or invalid credentials)"),
+            @ApiResponse(responseCode = "403", description = "Ops! You do not have permission to access this feature! :("),
+            @ApiResponse(responseCode = "404", description = "Resource not found"),
+            @ApiResponse(responseCode = "500", description = "Internal server error",
+                    content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDetails.class))})})
+    @ResponseStatus(value = HttpStatus.OK)
+    @GetMapping("/get-by-admin/{id}")
+    public ResponseEntity<List<Grupo>> getByAdmin(@PathVariable(value = "id") Integer id) {
+        return ok().body(service.findByAdmin(id));
+    }
+
+    @Operation(summary = "The list of grupos Pageable.", description = "Returns the list of entities.")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Request Ok"),
             @ApiResponse(responseCode = "401", description = "Not authenticated agent (missing or invalid credentials)"),
@@ -83,21 +99,22 @@ public class CursoController {
                     content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDetails.class))})})
     @ResponseStatus(value = HttpStatus.OK)
     @GetMapping("search")
-    public ResponseEntity<Page<Curso>> search(@RequestParam("searchTerm") String searchTerm,
-                                              @RequestParam(value = "order", required = false, defaultValue = "descricao") String order,
+    public ResponseEntity<Page<Grupo>> search(@RequestParam("searchTerm") String searchTerm,
+                                              @RequestParam(value = "userId", required = false) Integer userId,
+                                              @RequestParam(value = "order", required = false, defaultValue = "nome") String order,
                                               @RequestParam(value = "page", required = false, defaultValue = "0") int page,
                                               @RequestParam(value = "size", required = false, defaultValue = "10") int size) {
         String finalOrder = order.startsWith("-") ? order.substring(1) : order;
         SearchFilter searchFilter = Match(finalOrder.toLowerCase()).of(
-                Case($(isIn("id", "descricao")), new SearchFilter(searchTerm, order.toLowerCase(), page, size)),
+                Case($(isIn("id", "nome")), new SearchFilter(searchTerm, order.toLowerCase(), page, size)),
                 Case($(), () -> {
-                    throw new IllegalArgumentException("The sent sorting field is invalid. Available fields: 'id' and 'descricao.");
+                    throw new IllegalArgumentException("The sent sorting field is invalid. Available fields: 'id' and 'nome.");
                 })
         );
-        return ResponseEntity.ok(service.cursoPageable(searchFilter));
+        return ResponseEntity.ok(service.gruposPageable(userId, searchFilter));
     }
 
-    @Operation(summary = "Update cursos ", description = "Return Update cursos ")
+    @Operation(summary = "Update grupo", description = "Return Update grupo")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Request Ok"),
             @ApiResponse(responseCode = "401", description = "Not authenticated agent (missing or invalid credentials)"),
@@ -107,17 +124,19 @@ public class CursoController {
                     content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDetails.class))})})
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @PutMapping("{id}")
-    @PreAuthorize("hasRole('ROLE_MANAGER')")
-    public ResponseEntity<?> updateCurso(@PathVariable(value = "id") Integer id, @RequestBody @Valid Curso entity) {
+    public ResponseEntity<?> updateGrupo(@PathVariable(value = "id") Integer id, @RequestBody @Valid Grupo entity) {
         if (!Objects.equals(entity.getId(), id)) {
             return badRequest().body("Ops! Id of entity is not equals as param 'id'! :(");
+        } else if (!validaUsuarioAdmin(SecurityContextHolder.getContext(), id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Ops! You do not have permission to update this group!");
         } else {
             service.saveOrUpdate(entity);
             return noContent().build();
         }
     }
 
-    @Operation(summary = "Delete cursos", description = "Return Delete cursos by id (Long)")
+    @Operation(summary = "Delete grupos", description = "Return Delete grupos by id (Long)")
     @ApiResponses(value = {
             @ApiResponse(responseCode = "200", description = "Request Ok"),
             @ApiResponse(responseCode = "401", description = "Not authenticated agent (missing or invalid credentials)"),
@@ -127,9 +146,26 @@ public class CursoController {
                     content = {@Content(mediaType = "application/json", schema = @Schema(implementation = ErrorDetails.class))})})
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
     @DeleteMapping("{id}")
-    @PreAuthorize("hasRole('ROLE_MANAGER')")
     public ResponseEntity<?> delete(@PathVariable("id") Integer id) {
-        service.delete(service.findById(id));
-        return noContent().build();
+        if (!validaUsuarioAdmin(SecurityContextHolder.getContext(), id)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN)
+                    .body("Ops! You do not have permission to delete this group!");
+        } else {
+            service.delete(service.findById(id));
+            return noContent().build();
+        }
+    }
+
+    boolean validaUsuarioAdmin(SecurityContext securityContext, Integer idGrupo) {
+        // Recupera o usuario
+        Usuario usuario = (Usuario) securityContext.getAuthentication().getPrincipal();
+        // Recupera o grupo existente do banco de dados
+        Grupo existingGroup = service.findById(idGrupo);
+        // Verifica se o grupo existe
+        if (existingGroup == null) {
+            return false;
+        }
+        return existingGroup.getAdmin().getId().equals(usuario.getId());
     }
 }
+
